pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "mistralai>=1.9.10",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "requests>=2.32.5",
    "werkzeug>=3.1.3",
]
